frontend/pyproject.toml
[build-system]
requires = ['setuptools>=60', 'Cython==3.0.7', 'wheel==0.42.0']
build-backend = "setuptools.build_meta"

[project]
name = "ora-frontend"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi[standard]==0.115.14",
  "streamlit==1.40.2",
  "requests==2.32.4",
  "requests-oauthlib==2.0.0",
  "Pillow==11.0.0",
  "google-auth==2.30.0",
  "google-auth-httplib2==0.2.0",
  "google-auth-oauthlib==1.2.0",
  "gspread==6.1.2",
  "python-dotenv==1.0.1",
  "pymongo==4.6.3",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Software Development",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only",
]

[project.optional-dependencies]
test = [
  "pytz==2024.1",
  "mypy==1.10.1",
  "types-requests==2.32.0.20240622",
  "pre-commit==3.7.1",
  "ruff==0.5.1",
  "types-Flask-Cors==6.0.0.20250809",
]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
strict_optional = true
disable_error_code = ["call-arg"]
exclude = "src/post_install.py"

[[tool.mypy.overrides]]
module = ["flask.*", "pymongo.*", "gspread.*"]
ignore_missing_imports = true

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 88
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = ["E4", "E7", "E9","E301","E304","E305","E401","E223","E224","E242", "E", "F" ,"N", "W", "C90"]
extend-select = ["D204"]
ignore = ["E501", "C901"]
preview = true

# Allow fix for all enabled rules (when `--fix` is provided).
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"
